Move EAX files to their own sub-directory
