+++ al/eax/eax_call.cpp
@@ -1,8 +1,7 @@
 #include "config.h"
 
-#include "al/eax_eax_call.h"
-
-#include "al/eax_exception.h"
+#include "eax_call.h"
+#include "exception.h"
 
 
 namespace {
