+++ al/eax/eax_call.h
@@ -6,8 +6,8 @@
 
 #include "alspan.h"
 
-#include "eax_api.h"
-#include "eax_fx_slot_index.h"
+#include "api.h"
+#include "fx_slot_index.h"
 
 
 enum class EaxEaxCallPropertySetId
