+++ al/eax/effect.cpp
@@ -1,0 +1,3 @@
+#include "config.h"
+
+#include "effect.h"
