+++ al/eax/fx_slot_index.h
@@ -5,7 +5,7 @@
 #include <cstddef>
 
 #include "aloptional.h"
-#include "eax_api.h"
+#include "api.h"
 
 
 using EaxFxSlotIndexValue = std::size_t;
