+++ al/eax/globals.h
