+++ al/eax/utils.h
