--- al/eax_effect.cpp
@@ -1,3 +1,0 @@
-#include "config.h"
-
-#include "eax_effect.h"
