--- al/eax_x_ram.cpp
@@ -1,3 +1,0 @@
-#include "config.h"
-
-#include "eax_x_ram.h"
