+++ al/effects/autowah.cpp
@@ -14,8 +14,8 @@
 #ifdef ALSOFT_EAX
 #include "alnumeric.h"
 
-#include "al/eax_exception.h"
-#include "al/eax_utils.h"
+#include "al/eax/exception.h"
+#include "al/eax/utils.h"
 #endif // ALSOFT_EAX
