+++ al/effects/distortion.cpp
@@ -10,8 +10,8 @@
 #ifdef ALSOFT_EAX
 #include "alnumeric.h"
 
-#include "al/eax_exception.h"
-#include "al/eax_utils.h"
+#include "al/eax/exception.h"
+#include "al/eax/utils.h"
 #endif // ALSOFT_EAX
