+++ al/effects/reverb.cpp
@@ -11,10 +11,13 @@
 
 #ifdef ALSOFT_EAX
 #include <tuple>
-#include "alnumeric.h"
+
 #include "AL/efx-presets.h"
-#include "al/eax_exception.h"
-#include "al/eax_utils.h"
+
+#include "alnumeric.h"
+
+#include "al/eax/exception.h"
+#include "al/eax/utils.h"
 #endif // ALSOFT_EAX
