+++ alc/device.h
@@ -20,7 +20,7 @@
 #include "vector.h"
 
 #ifdef ALSOFT_EAX
-#include "al/eax_x_ram.h"
+#include "al/eax/x_ram.h"
 #endif // ALSOFT_EAX
 
 struct ALbuffer;
